requirements.txt
pytest==6.2.2
selenium==3.141.0
allure-pytest=2.9.43